Add a new signature module: "reads_self"
Reports if any binary running in the sandbox opens and reads itself by keeping track of handles across each process.  It reports the offset and length of each read as well so an analyst can dig in immediately to those offsets to look for embedded binaries or configuration data.
Requires several recent improvements to cuckoomon and cuckoo to operate: the raw_value addition to arguments, the storage of the module's full normalized pathname, the addition of Length to NtReadFile, etc.