Create ek_heapsray.py
Attempt to detect JavaScript obfuscated exploits containing common heap spray values which sometimes appears. Memory locations taken from EMET for the memory locations it looks to for heapspray protections as being commonly attacked although obviously there is many other ways to accomplish this in an exploit that this is low level fruit attempt at detection.

If features from cuckoosploit in the future will later be added in the future this would be more effective for its detection of heapsprays, ROP chains etc.