Update Hawkeye signature
Restructure the signature a bit and add email terms to a list variable, iterating it at the time of the send api. Also decode base64 buffers as new Hawkeye samples were observed to be using base64 instead of plain text. Add in an additional email term observed with new samples, as well as a new IOC.