Create powershell_command.py
Check for different parameters. Generally the bad one is bypass of execution policy and to a lesser extent not to load current user profile but i have created it to provide some more information on suspicious or interesting actions (or at least being able to add them) with a severity of 2 and then if a security bypass is attempted (executionpolicy/noprofile) then the severity is changed to 3.