Add new detections into office_macro
Add in detections for known malicious authors & edit time as well as modifications to append date. This means for an office document with macros the main message will always be "The office file has X macros" as either a severity 2 or 3 depending if this is modified. 

I did want to do self.description += " The file also appears to contain suspicious characteristics." so that when either of these detected it would tag that onto the main message. However I was uncertain at the time how to make it so that it would only append this to the end once even if multiple hits fired.