Update java static file with new detections
Added in more exploit and generic detections, put adding on the severity increase & appending of the description until the end of the exploit checks to save on repeating lines from any exploit added in. Also on the modifications to security and permissions modified it to check reversed strings as well as various hex representations of those strings.