Update office macro sig
Add in detections where edit time instead of being 0 is None, same also for number of words. Also added in a new detection for the same thing with number of pages. 

This has been seen in recent Dridex emails (i.e MD5 07010709d3a84b41e1367adb3ea3dac8) where only parts that fire from this was that it had 4 macros and it was a known malicious auto-generated author ("1").