Add signature to detect common HTTP CnC features
This is the start of a signature to try and pick out various anomalies. In large volumes traffic many of these features will be highly common but should be interesting from a sandbox environment where only anomalous or cnc related traffic. 

I will add in other general features in time and then I intent to look at best ways to present likely CnC hosts (appending the HTTP hostname of the most likely cnc candidates to the alerts).