Create static_pdf analysis document
This brings together some of the existing static PDF analysis documents so if this is accepted the files pdf_eof.py & pdf_page.py should be removed.

The idea behind this is along the same lines as the static_java signature where indicators and exploit detections will be added in. For now I have gor the general flow of the document going, transfered existing checks which can be brought under this and added in a few more checks too. I also intend to add in exploit checks both specific and generic (i.e heap spray code etc) including older exploits as they are useful both for learning & in case of examples against unprotected systems still with vulnerable versions.