Create signature for doc file initiating network comms
Create a signature to detect when a document file (Office/PDF) initiates network communications. I whitelist may need to be added at a later date depending on FPs but my sandbox has not produced any. 

The aim of the signature is to detect exploit payloads or other features - especially in a closed sandbox environment where other malicious features may not be revealed by signatures and so this helps makes these connections more visible as a document should not make random connections when executed in a sandbox under normal conditions.  This has been tested against old and new Dridex doc samples and MWI document exploits.