Add a file for EK related generated redirects
This looks for a JS generated style element with an iframe redirect. I've burned in for a few weeks now, primarily firing on Angler. However it was observed today to also be firing on Neutrino. Thus I'm committing this with only the one IOC. Other indicators will be added in at a later time.