Add in new author noted in Dridex campaigns
Added in another author seen in Dridex campaigns. Also is the network document traffic signature that has not been merged ok or is there work that needs done? I have found it never useful in macro campaigns if analysing in a closed sandbox and the EXE is unavailable to make it available to others the download paths to track.