Add in extraction for GET no UA
Add in CnC score to GET no user agent indicator. This was not present due to other use of the scoring in early iterations. As more useful specific indicators are added this may be revised or utilised differently but right now it is necessary to extract and display the request as appended data also.