Create ek_cve2015_2419.py
Add in specific signature for this exploit. The idea will be to eventually collapse some sigs into a sig with indicators and specific exploits defined for IE and things (ones being exploited in wild though only because don't want to go crazy; just provide extra info). Tried doing this already but did not work but didn't have long to look at why when defining this check and trying to get the heaps spray sig cleaned up alongside it.

Just providing this though the now given it is still being exploited by some EKs so may be useful at least for a while.