Add in new generic dropped files technique
Add in a new technique to detect ransomware regardless of API calls used (moved files, create files etc) that ransomware may use. I simply check if a large number of files have been dropped with the type being "data" which means it is unrecognised. This can be tuned further later to be a bit more accurate and also to integrate it into the appended file extensions if a large number of files have same extension appended to double file extension (a change for later).

For now this should provide a useful indicator of ransomware regardless of the family.