Fix traceback
This traceback happens when we don't observe any winapi to NtCreateNamedPipeFile. Since we are accounting for child processes to create the pipe, we need to check to see it the creation happened in a process we saw. If it didn't (EG some service management/IE backending) this would cause a traceback as we didn't see the creation event, but saw the NtWriteFile to the service pipe.