Update ek_virtualcheck sig
Added in 2 new indicators, a fix by making the indicator .lower() too and also doing a removal of the following characters:

- A space
- '
- "
- +

This results in a messy but effective removal of any split string obfuscation for a second check to see if the indicators exist in the remaining buffer. Tested against recent Angler landing pages and it works again.