Add signature to detect dropping ransom messages
First version of a signature to detect files writing likely ransomware messages to disk. Currently once I have it I just output the filename; the later intention is to extract the filename itself rather than the whole path but I was having difficulty getting it to work right. 

That would be preferable as ransomware messages typically are the same file dropped throughout the filesystem so just getting the file will take the list down from dozens to likely just one once it is working but I will look at sorting that up another time as this signature works for now and tried against a few ransomware families.