Handle appended email addresses differently
When appending files some ransomware families append email addresses to the encrypted files as well as showing this in a ransom message. This results in innacurate reporting for the appended extension extraction which usually then just ends up extracting the TLD from the email rather than an extension.

So to handle this we don't bother extracting extension if likely an email appended (@ in extension) and instead just provide a message that this has been done.